fix integer types in iterative matrix methods
I noticed that gauss_seidel in matrixSolve.py has a bug in handling int and float data types:

For gauss_seidel, run it on the following system:

A = np.array([[2,1,0],[1,3,1],[0,1,2]])

b = np.array([[6],[10],[6]])

u = np.array([[1],[2],[3]])

It will give a wrong answer as [[2],[1],[2]], where the correct answer should be [[2],[1.667],[2.1667]]. The reason is:

In `u[j] = u[j] + (b[j] - np.dot(A[j, :], u)) / A[j, j]`, although the right hand side might be non-integer, e.g., 1.667, the left hand side is of type int64, which will truncate the float number
to an integer, giving the wrong result.

To fix the bug, add the following after A = A.astype(float):

u = u.astype(float)

b = b.astype(float)

+++ code/matrixSolve.py
@@ -259,8 +259,10 @@ def gauss_seidel(A, u, b, n_iterations):
     # Get dimension
     k = len(A)
 
-    # Make sure matrix A is float
+    # Make sure we are working with floats
     A = A.astype(float)
+    u = u.astype(float)
+    b = b.astype(float)
 
     for i in range(n_iterations):
         for j in range(k):
@@ -286,8 +288,10 @@ def jacobi(A, u, b, n_iterations):
     # Get dimension
     k = len(A)
 
-    # Make sure matrix A is float
+    # Make sure we are working with floats
     A = A.astype(float)
+    u = u.astype(float)
+    b = b.astype(float)
 
     for i in range(n_iterations):
         r = b - np.dot(A, u)
@@ -315,8 +319,10 @@ def jacobi2(A, u, b, n_iterations):
     # Get dimension
     k = len(A)
 
-    # Make sure matrix A is float
+    # Make sure we are working with floats
     A = A.astype(float)
+    u = u.astype(float)
+    b = b.astype(float)
 
     unew = np.zeros([k, 1])
     for i in range(n_iterations):
@@ -343,8 +349,10 @@ def gauss_seidel_new(A, u, b, tol):
     # Get dimension
     k = len(A)
 
-    # Make sure matrix A is float
+    # Make sure we are working with floats
     A = A.astype(float)
+    u = u.astype(float)
+    b = b.astype(float)
 
     # Set the maximum number of iterations
     maxit = 1000
@@ -387,8 +395,10 @@ def jacobi_new(A, u, b, tol):
     # Get dimension
     k = len(A)
 
-    # Make sure matrix A is float
+    # Make sure we are working with floats
     A = A.astype(float)
+    u = u.astype(float)
+    b = b.astype(float)
 
     # Set the maximum number of iterations
     maxit = 1000
